requirements.txt
alembic==1.12.1
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.3
cachelib==0.10.2
charset-normalizer==3.1.0
click==8.1.3
dateutils==0.6.12
dnspython==2.3.0
email-validator==1.3.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.9
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mongoengine==0.27.0
mongoengine-goodjson==1.1.8
Pillow==9.4.0
PyJWT==2.7.0
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==1.26.15
Werkzeug==3.0.1
WTForms==3.0.1
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
